requirements.txt
uvicorn==0.19.0
starlette==0.21.0
#pymongo==4.3.2
#pydantic==1.10.2
#fastapi_jwt_auth~=0.5.0
python-decouple~=3.5
dnspython~=2.2.0
requests==2.28.1
humanize==4.4.0
coloredlogs~=15.0.1
verboselogs~=1.7
#backoff==2.2.1
fastapi==0.85.1
PyJWT==2.6.0
